Move files to new directory. This makes the directory structure compliant to the python package layout.
